packages/backend/notification-service/requirements.txt
fastapi
uvicorn
paho-mqtt==2.1.0
redis
